Delay game start to after camera slide in animation
